[Artemis] Update macros to fix heightmap warning
Homing uses the endstops to set the Z-height, but probing uses the
strain gauge to set the same. If you create a heightmap using the probe,
but then home against the endstops your heightmap could be offset from
the homed z=0 reference point (datum).

This now causes RRF to issue a warning:
> Warning: the height map was loaded when the current Z=0 datum was not
> determined probing. This may result in a height offset.

I think replacing G28 commands issued after probing with a custom macro
that moves the effector out of the way should avoid the warnings.

I also took this opportunity to modernize my start-print macro as well.