[ARTEMIS] Add pressure advance test and macro
gcodes/marlin_linadv_test.gcode is a test print to help dial in pressure
advance (currently configured for what I found for Atomic's Silky Lilac PLA)

Created a new macro directory for Pressure Advance with a macro to set
it for the Silky Lilac PLA; as we configure PA for each filament, add a
macro for it here.

TODO: make it so we can pick which extruder to set the PA for?